requirements.txt
# Core Flask
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Marshmallow==0.15.0
marshmallow-sqlalchemy==0.29.0

# Caching
Flask-Caching==2.1.0
redis==5.0.1

# CORS
Flask-CORS==4.0.0

# Rate Limiting
Flask-Limiter==3.5.0

# Environment Variables
python-dotenv==1.0.0

# Database - Version compatible Python 3.14
SQLAlchemy==2.0.36
alembic==1.14.0

# Utilities
Werkzeug==3.0.1
requests==2.31.0

# Testing
pytest==7.4.3
pytest-flask==1.3.0
pytest-cov==4.1.0

# Validation
marshmallow==3.20.1

# Production Server
gunicorn==21.2.0
psycopg2-binary==2.9.9
